Michael branch

+++ server/src/controllers/auth.ts
@@ -72,6 +72,10 @@ export const sendReVerificationToken: RequestHandler = async (req, res) => {
 
   if (!user) return res.status(403).json({ error: "Invalid request!" });
 
+  // User is already verified
+  if (user.verified)
+    return res.status(422).json({ error: "Your account is already verified!" });
+
   await EmailVerificationToken.findOneAndDelete({
     owner: userId,
   });
